+++ packages/malloy-render/src/html/sparkline.ts
